do not assume time_t is long
On Linux time_t is long, on OpenBSD time_t is long long... so printf
statements using %lu on 32-bit platforms will break on !Linux.
The only portable way to print a time_t is using a cast and casting to
"long long" (gint64) is probably the most portable way.

https://bugzilla.gnome.org/show_bug.cgi?id=707900

+++ addressbook/backends/file/e-book-backend-file.c
@@ -613,7 +613,7 @@ e_book_backend_file_create_unique_id (void)
 	 * it's doubtful 2^32 id's will be created in a second, so we
 	 * should be okay. */
 	static guint c = 0;
-	return g_strdup_printf (PAS_ID_PREFIX "%08lX%08X", time (NULL), c++);
+	return g_strdup_printf (PAS_ID_PREFIX "%08lX%08X", (glong) time (NULL), c++);
 }
 
 static gchar *
